==> notebooks/01_enedis_preparation.ipynb <==
import pandas as pd
import numpy as np
import unicodedata

file_path = "enedis_raw.csv"

df_raw = pd.read_csv(file_path, sep=";", low_memory=False)

df_raw.head()

df_raw.shape

On va garder les colonnes qui semblent les plus utiles

cols = [
    "annee",
    "numero_de_voie",
    #"indice_de_repetition",
    "type_de_voie",
    "libelle_de_voie",
    "code_commune",
    "nom_commune",
    "nombre_de_logements",
    "consommation_annuelle_totale_de_l_adresse_mwh",
    "adresse",
    "code_departement",
    "code_region"
]

df = df_raw[cols].copy()

df.head()

df.shape

Le calcul de la consommation par logement en kWh/an

# Passage en numérique
df["nombre_de_logements"] = pd.to_numeric(df["nombre_de_logements"], errors="coerce")
df["consommation_annuelle_totale_de_l_adresse_mwh"] = pd.to_numeric(df["consommation_annuelle_totale_de_l_adresse_mwh"], errors="coerce")

# Au cas ou valeur a 0 ou NaN
mask_valid = df["nombre_de_logements"] > 0

# La division de la conso annuelle de l'adresse par son nombre de logement -> conso par logement
df["conso_par_logement_MWh"] = np.where(mask_valid,df["consommation_annuelle_totale_de_l_adresse_mwh"] / df["nombre_de_logements"],np.nan)

# Conversion en kWh/an
df["conso_par_logement_kWh"] = df["conso_par_logement_MWh"] * 1000

df[["nombre_de_logements", "consommation_annuelle_totale_de_l_adresse_mwh", "conso_par_logement_MWh", "conso_par_logement_kWh"]].head()

On essaie de formaliser les adresses dans un style : 

"numero_de_voie + type_de_voie + libelle_de_voie + nom_commune"

def normalize_text(s: str) -> str:
    if pd.isna(s):
        return ""
    # passage en str
    s = str(s)
    # enlever accents
    s = unicodedata.normalize('NFKD', s)
    s = "".join(c for c in s if not unicodedata.combining(c))
    # majuscules
    s = s.upper()
    # remplace toute ponctuation par espace
    for ch in [",", ";", ":", ".", "/", "\\", "'", '"', "(", ")", "-", "_"]:
        s = s.replace(ch, " ")
    # espaces en trop
    s = " ".join(s.split())
    return s

# Nettoyage du numéro de voie
df["numero_de_voie_num"] = (pd.to_numeric(df["numero_de_voie"], errors="coerce").round().astype("Int64"))

# On transforme en string propre pour la concaténation
num = df["numero_de_voie_num"].astype("object").fillna("").astype(str)

# Nettoyage du type de voie 
df["type_de_voie"] = df["type_de_voie"].astype(str)
df["type_de_voie"] = df["type_de_voie"].replace(["0", "0.0", "nan", "None", "NONE", "NaN"], "")
typ = df["type_de_voie"].fillna("").astype(str)

# Nettoyage du libelle de voie
nom_voie = df["libelle_de_voie"].fillna("").astype(str)

# Nettoyage du nom de commune
commune = df["nom_commune"].fillna("").astype(str)

# Construction
adresse_brut_concat = num + " " + typ + " " + nom_voie + " " + commune

# Normalisation 
df["adresse_clean"] = adresse_brut_concat.apply(normalize_text)

df[["adresse", "adresse_clean"]].head(10)

nb_diff = (df["adresse"].apply(normalize_text) != df["adresse_clean"]).sum()
pct_diff = nb_diff / len(df) * 100
pct_diff
print("Nombre d'adresse modifié : ", nb_diff, "soit", pct_diff, "%") 

df.head()

df.to_csv("enedis_clean.csv", index=False, encoding="utf-8")
