==> notebooks/04_Comparison.ipynb <==
import pandas as pd
import numpy as np

enedis = pd.read_csv("enedis_clean.csv")
ban = pd.read_csv("enedis_ban_final.csv")

enedis.columns

ban.columns

ban

enedis

cols_enedis = [
    "annee",
    "numero_de_voie",
    "type_de_voie",
    "libelle_de_voie",
    "code_commune",
    "nom_commune",
    "nombre_de_logements",
    "consommation_annuelle_totale_de_l_adresse_mwh",
    "adresse",
    "code_departement",
    "code_region",
    "conso_par_logement_MWh",
    "conso_par_logement_kWh",
    "numero_de_voie_num",
    "adresse_clean",
]

# On garde seulement les colonnes utiles côté Enedis
enedis_sub = enedis[cols_enedis].copy()

# Jointure : toutes les lignes BAN + colonnes Enedis
ban_enrichi = ban.merge(
    enedis_sub,
    on="adresse_clean",
    how="left",
    suffixes=("_ban", "_enedis")
)

nb_ban = len(ban_enrichi)
nb_match = ban_enrichi["annee"].notna().sum()  # ou une autre colonne Enedis non nulle
print(f"Total lignes BAN           : {nb_ban}")
print(f"Lignes avec match Enedis   : {nb_match}")
print(f"Lignes BAN sans match      : {nb_ban - nb_match}")

ban_enrichi.columns

addr = "Première Avenue 06600 Antibes"
ban_enrichi[ban_enrichi["adress_ban"] == addr].sort_values("annee")

ban_enrichi = ban_enrichi.drop(columns=["adress_ban_clean"])

ban_enrichi

ban_enrichi.to_csv("enedis_clean_final.csv", index=False)

import pandas as pd
import numpy as np

test = pd.read_csv("enedis_clean_final.csv")

test[test["adress_ban"] == "89 Rue du Marais 59160 Lille"]

df_clean = test.drop_duplicates()

nb_avant = len(test)
nb_apres = len(df_clean)
print("Doublons supprimés :", nb_avant - nb_apres)
